refactor the packages for the notification data service
