+++ src/main/java/com/blackducksoftware/integration/hub/dataservices/notifications/NotificationDataService.java
@@ -1,4 +1,4 @@
-package com.blackducksoftware.integration.hub.dataservices;
+package com.blackducksoftware.integration.hub.dataservices.notifications;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -24,16 +24,17 @@ import com.blackducksoftware.integration.hub.api.notification.NotificationItem;
 import com.blackducksoftware.integration.hub.api.notification.PolicyOverrideNotificationItem;
 import com.blackducksoftware.integration.hub.api.notification.RuleViolationNotificationItem;
 import com.blackducksoftware.integration.hub.api.notification.VulnerabilityNotificationItem;
-import com.blackducksoftware.integration.hub.dataservices.items.AbstractItemCount;
-import com.blackducksoftware.integration.hub.dataservices.items.NotificationContentItem;
-import com.blackducksoftware.integration.hub.dataservices.items.NotificationCountData;
-import com.blackducksoftware.integration.hub.dataservices.items.NotificationItemCount;
-import com.blackducksoftware.integration.hub.dataservices.items.PolicyNotificationFilter;
-import com.blackducksoftware.integration.hub.dataservices.items.VulnerabilityItemCount;
-import com.blackducksoftware.integration.hub.dataservices.transforms.AbstractNotificationTransform;
-import com.blackducksoftware.integration.hub.dataservices.transforms.PolicyViolationOverrideTransform;
-import com.blackducksoftware.integration.hub.dataservices.transforms.PolicyViolationTransform;
-import com.blackducksoftware.integration.hub.dataservices.transforms.VulnerabilityTransform;
+import com.blackducksoftware.integration.hub.dataservices.AbstractDataService;
+import com.blackducksoftware.integration.hub.dataservices.notifications.items.AbstractItemCount;
+import com.blackducksoftware.integration.hub.dataservices.notifications.items.NotificationContentItem;
+import com.blackducksoftware.integration.hub.dataservices.notifications.items.NotificationCountData;
+import com.blackducksoftware.integration.hub.dataservices.notifications.items.NotificationItemCount;
+import com.blackducksoftware.integration.hub.dataservices.notifications.items.PolicyNotificationFilter;
+import com.blackducksoftware.integration.hub.dataservices.notifications.items.VulnerabilityItemCount;
+import com.blackducksoftware.integration.hub.dataservices.notifications.transforms.AbstractNotificationTransform;
+import com.blackducksoftware.integration.hub.dataservices.notifications.transforms.PolicyViolationOverrideTransform;
+import com.blackducksoftware.integration.hub.dataservices.notifications.transforms.PolicyViolationTransform;
+import com.blackducksoftware.integration.hub.dataservices.notifications.transforms.VulnerabilityTransform;
 import com.blackducksoftware.integration.hub.exception.BDRestException;
 import com.blackducksoftware.integration.hub.rest.RestConnection;
 import com.google.gson.Gson;
