+++ src/test/java/com/blackducksoftware/integration/hub/dataservices/items/VulnerabilityItemCountTest.java
@@ -8,6 +8,7 @@ import org.mockito.Mockito;
 
 import com.blackducksoftware.integration.hub.api.notification.VulnerabilityNotificationContent;
 import com.blackducksoftware.integration.hub.api.notification.VulnerabilityNotificationItem;
+import com.blackducksoftware.integration.hub.dataservices.notifications.items.VulnerabilityItemCount;
 
 public class VulnerabilityItemCountTest {
